Add cryptid by name refactor
