+++ app/graphql/types/query_type.rb
@@ -5,19 +5,9 @@ module Types
     include GraphQL::Types::Relay::HasNodesField
 
     field :cryptid_by_id, resolver: Queries::CryptidById
-    # field :cryptid_by_name, resolver: Queries::CryptidByName
-    # field :cryptid_by_range, resolver: Queries::CryptidByRange
+    field :cryptid_by_name, resolver: Queries::CryptidByName
     field :get_cryptids, resolver: Queries::GetCryptids
-
-
-    field :cryptid_by_name, CryptidType, "Cryptid by name" do
-      argument :name, String
-    end
-
-    def cryptid_by_name(args)
-      Cryptid.find_by(name: args[:name])
-    end
-
+    # field :cryptid_by_range, resolver: Queries::CryptidByRange
 
     field :sightings, [SightingType], null: false,
       description: "All sightings"
